chalice/redshift_to_s3/requirements.txt
psycopg2-binary
boto3
func_timeout
pandas